style:  warnings de compilador solucionadas

+++ src/headers/filehandler_parallel.h
@@ -109,10 +109,10 @@ int targetdir_compress_parallel(TargetDir* td, char **dict){
         char bit;
         
         // por cada caracter
-        for(int j=0; j<td->file_sizes[i]; j++){
+        for(size_t j=0; j<td->file_sizes[i]; j++){
             int caracter = td->content[i][j];
             //por cada bit de codigo
-            for(int k=0; k<strlen(dict[caracter]); k++){
+            for(size_t k=0; k<strlen(dict[caracter]); k++){
                 char bitchar = dict[caracter][k];
                 //Insertar el bit
                 if(bitchar=='0'){
